Migrate to Tauri v2.x
(by running `cargo tauri migrate` as per "https://v2.tauri.app/start/migrate/from-tauri-1/")
