+++ crates/tauri/capabilities/desktop.json
@@ -1,0 +1,17 @@
+{
+  "identifier": "desktop-capability",
+  "platforms": [
+    "macOS",
+    "windows",
+    "linux"
+  ],
+  "windows": [
+    "main"
+  ],
+  "permissions": [
+    "global-shortcut:default",
+    "http:default",
+    "notification:default",
+    "process:default"
+  ]
+}
