+++ crates/tauri/src/lib.rs
@@ -1,0 +1,34 @@
+#![cfg_attr(
+    all(not(debug_assertions), target_os = "windows"),
+    windows_subsystem = "windows"
+)]
+
+use backend::app;
+
+struct Port(u16);
+
+#[cfg_attr(mobile, tauri::mobile_entry_point)]
+pub fn run() {
+    let port = portpicker::pick_unused_port().expect("failed to find unused port");
+    tauri::async_runtime::spawn(app(port));
+    tauri::Builder::default()
+        .plugin(tauri_plugin_process::init())
+        .plugin(tauri_plugin_notification::init())
+        .plugin(tauri_plugin_http::init())
+        .plugin(tauri_plugin_global_shortcut::Builder::new().build())
+        .plugin(tauri_plugin_os::init())
+        .plugin(tauri_plugin_dialog::init())
+        .plugin(tauri_plugin_clipboard_manager::init())
+        .plugin(tauri_plugin_shell::init())
+        .plugin(tauri_plugin_fs::init())
+        .manage(Port(port))
+        .invoke_handler(tauri::generate_handler![get_port])
+        .run(tauri::generate_context!())
+        .expect("error while running tauri application");
+}
+
+/// A command to get the usused port, instead of 3000.
+#[tauri::command]
+fn get_port(port: tauri::State<Port>) -> Result<String, String> {
+    Ok(format!("{}", port.0))
+}
